+++ PharmVerse/PharmVerse.Domain/Events/Patients/PatientAccessedEvent.cs
@@ -1,0 +1,16 @@
+﻿using MediatR;
+using PharmVerse.Domain.Patients;
+
+namespace PharmVerse.Domain.Events.Patients
+{
+     public class PatientAccessedEvent : INotification
+     {
+          public PatientAccessedEvent(Patient patient)
+          {
+               Patient = patient;
+          }
+
+          public Patient Patient { get; }
+
+     }
+}
