--- PharmVerse/PharmVerse.Domain/Patients/Patients/PatientUpdatedEvent.cs
@@ -1,19 +1,0 @@
-﻿using MediatR;
-using PharmVerse.Domain.Patients;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PharmVerse.Domain.Patients.Patients
-{
-    public class PatientUpdatedEvent : INotification
-    {
-        public PatientUpdatedEvent(Patient patient)
-        {
-            Patient = patient;
-        }
-        public Patient Patient { get; }
-    }
-}
