+++ PharmVerse/PharmVerse.Domain/Users/AppUser.cs
@@ -12,14 +12,6 @@ namespace PharmVerse.Domain.Users
 {
      public class AppUser : IdentityUser, IHasDomainEvent
      {
-          public AppUser()
-          {
-               Invites = new HashSet<Invite>();
-               Notifications = new HashSet<Notification>();
-               Reviews = new HashSet<Review>();
-               Transactions = new HashSet<Transaction>();
-          }
-
           public string Title { get; set; }
           public string FirstName { get; set; }
           public string MiddleName { get; set; }
@@ -45,9 +37,9 @@ namespace PharmVerse.Domain.Users
           public virtual Wallet Wallet { get; set; }
 
           // One AppUser to Many Relationship
-          public virtual ICollection<Invite> Invites { get; set; }
-          public virtual ICollection<Notification> Notifications { get; set; }
-          public virtual ICollection<Review> Reviews { get; set; }
-          public virtual ICollection<Transaction> Transactions { get; set; }
+          public virtual ICollection<Invite> Invites { get; set; } = new HashSet<Invite>();
+          public virtual ICollection<Notification> Notifications { get; set; } = new HashSet<Notification>();
+          public virtual ICollection<Review> Reviews { get; set; } = new HashSet<Review>();
+          public virtual ICollection<Transaction> Transactions { get; set; } = new HashSet<Transaction>();
      }
 }
