+++ PharmVerse/PharmVerse.Presentation/Patients/Commands/DeletePatientCommand.cs
@@ -1,15 +1,11 @@
 ﻿using MediatR;
 using PharmVerse.Presentation.Patients.Models.Result;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PharmVerse.Presentation.Patients.Commands
 {
-    public class DeletePatientCommand : IRequest<DeletePatientCommandResult>
-    {
-        public Guid Id { get; set; }
-    }
+     public class DeletePatientCommand : IRequest<DeletePatientCommandResult>
+     {
+          public Guid Id { get; set; }
+     }
 }
