+++ PharmVerse/PharmVerse.Presentation/Patients/Models/Result/CreatePatientCommandResult.cs
@@ -1,25 +1,21 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PharmVerse.Presentation.Patients.Models.Result
 {
-    public class CreatePatientCommandResult
-    {
-        public Guid Id { get; set; }
-        public DateTime Date { get; set; }
-        public string FirstName { get; set; }
-        public string MiddleName { get; set; }
-        public string LastName { get; set; }
-        public string Image { get; set; }
-        public string Address { get; set; }
-        public string City { get; set; }
-        public string State { get; set; }
-        public string Country { get; set; }
-        public string DoctorInCharge { get; set; }
-        public string Ailment { get; set; }
-        
-    }
+     public class CreatePatientCommandResult
+     {
+          public Guid Id { get; set; }
+          public DateTime Date { get; set; }
+          public string FirstName { get; set; }
+          public string MiddleName { get; set; }
+          public string LastName { get; set; }
+          public string Image { get; set; }
+          public string Address { get; set; }
+          public string City { get; set; }
+          public string State { get; set; }
+          public string Country { get; set; }
+          public string DoctorInCharge { get; set; }
+          public string Ailment { get; set; }
+
+     }
 }
