+++ PharmVerse/PharmVerse.Presentation/Patients/Queries/GetPatientsForUserQuery.cs
@@ -1,16 +1,12 @@
 ﻿using MediatR;
 using PharmVerse.Presentation.Patients.Models.Result;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PharmVerse.Presentation.Patients.Queries
 {
-    public class GetPatientsForUserQuery : IRequest<IEnumerable<GetPatientQueryResult>>
-    {
-        public string AppUserId { get; set; }
+     public class GetPatientsForUserQuery : IRequest<IEnumerable<GetPatientQueryResult>>
+     {
+          public string AppUserId { get; set; }
 
-    }
+     }
 }
